+++ src/llms/openai/generate.ts
@@ -1,6 +1,7 @@
 import { Notice } from "obsidian";
 import OpenAI from "openai";
 import { LlmStreamingResponseFn, SimplePromptPluginSettings } from "src/types";
+import { notice } from "src/utils";
 
 export async function generate(
     settings: SimplePromptPluginSettings,
@@ -8,18 +9,18 @@ export async function generate(
     onSuccess: (result: string) => void,
 ) {
     const openai = new OpenAI({
-        apiKey: settings.apiKey ?? "",
+        apiKey: settings.apiKey.openai ?? "",
         dangerouslyAllowBrowser: true,
     });
 
     const response = await openai.chat.completions
         .create({
-            model: settings.model,
+            model: settings.model.openai,
             messages: [{ role: "user", content: prompt }],
         })
         .catch((e) => {
             console.error(e);
-            new Notice(`Error generating content: ${e}`);
+            notice(`Error generating content: ${e}`);
             return null;
         });
 
@@ -36,19 +37,19 @@ export async function generateStreaming(
     onEnd?: () => void,
 ) {
     const openai = new OpenAI({
-        apiKey: settings.apiKey ?? "",
+        apiKey: settings.apiKey.openai ?? "",
         dangerouslyAllowBrowser: true,
     });
 
     const stream = await openai.chat.completions
         .create({
-            model: settings.model,
+            model: settings.model.openai,
             messages: [{ role: "user", content: prompt }],
             stream: true,
         })
         .catch((e) => {
             console.error(e);
-            new Notice(`Error generating content: ${e}`);
+            notice(`Error generating content: ${e}`);
             return null;
         });
     if (stream) {