+++ src/utils.ts
@@ -1,0 +1,5 @@
+import { Notice } from "obsidian";
+
+export function notice(message: string) {
+    new Notice(`[Simple Prompt] ${message}`);
+}